+++ src/Solver/MatrixStorageFormats/CCSFormat.h
@@ -26,10 +26,10 @@ struct CCSFormat_t; //typedef struct CCSFormat_t      CCSFormat_t ;
  * colptr points to the columns in both rowind and nzval
  * If a_ij = nzval[k] then rowind[k] = i and colptr[j] <= k < colptr[j + 1] */
 struct CCSFormat_t {
-  unsigned int    nnz ;       /* nb of non zero values */
+  size_t    nnz ;       /* nb of non zero values */
   double* nzval ;             /* Non zero values */
-  unsigned int* rowind ;      /* Row indices of the non zeros */
-  unsigned int* colptr ;      /* Index of element in nzval which starts a column */
+  size_t* rowind ;      /* Row indices of the non zeros */
+  size_t* colptr ;      /* Index of element in nzval which starts a column */
 } ;
 
 #endif
